fix: 综合代码审计 v5 - 修复 68 项问题（6 P0 + 15 P1 + 8 P2 + 5 P3）
P0: ESP-NOW拓扑修复、DRIVE_MODE协议重构('T')、串口解析器重构、WS自动连接、视频缓冲区32KB、紧急停止显式解除
P1: 航向归一化、MAC配对更新、HEADING_LOCK实现、dotenv时序、命令失败感知、串口重启、MAC帧边界、build.rs修复
P2: Base64共享编码、/api/ports缓存、增量平均速度、校准系数约束、移除Pinia/死代码
P3: appearance修复、箭头键修复、duty溢出修复、volatile清理

+++ firmware/car_controller/servo_control.h
@@ -246,7 +246,8 @@ inline void initializeServoPins() {
 inline void applyServoState(const ServoState& state, const ServoConfig& config) {
     const uint16_t pulse = angleToPulse(state.currentAngle, config);
     // 将微秒转换为duty cycle（16位分辨率）
-    const uint32_t duty = (pulse * 65535ULL) / 20000; // 20ms周期
+    // 使用 uint64_t 中间计算防止乘法溢出（pulse 最大 2400，2400*65535 = 157M，超出 uint32_t 范围）
+    const uint32_t duty = static_cast<uint32_t>((static_cast<uint64_t>(pulse) * 65535ULL) / 20000ULL);
     ledcWrite(config.pin, duty);
 }
